pyproject.toml
[project]
name = "fabrica-hackathon"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "fabricaDeSoftware", email = ""},
]
dependencies = [
    "django>=5.0.4",
    "pillow>=10.3.0",
    "django-rest-framework>=0.1.0",
    "django-cors-headers>=4.4.0",
    "drf-spectacular>=0.27.2",
    "djangorestframework-simplejwt>=5.3.1",
    "celery>=5.4.0",
    "django-celery-results>=2.5.1",
    "dotenv>=0.0.5",
    "python-dotenv>=1.0.1",
    "django-filter>=24.3",
    "django-base64field>=1.0",
    "requests>=2.32.3",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
pre_migrate = "python manage.py makemigrations"
migrate = "python manage.py migrate"
shell = "python manage.py shell"
shellp= "python manage.py shell_plus"
test = "python manage.py test"
populate = "python manage.py populate"
loaddata = "python manage.py loaddata core.json"
dumpdata = "python manage.py dumpdata --indent 2"
celery = "celery -A config worker -l info"
flush = "python manage.py flush"

[tool.pdm]
distribution = false
